+++ CHANGES.md
@@ -1,3 +1,8 @@
+# 1.3 (unreleased)
+
+* Improve regexp for extracting the return type of c-family functions from
+  function signature
+
 # 1.2 (Jun 14, 2017)
 
 * New `ycmd-eldoc-mode`.